refactor(channel_finder): decompose hierarchical database class
HierarchicalChannelDatabase becomes a thin composition of six
concern-scoped mixins (base helpers, loading/validation, naming,
query, preview, write) with an unchanged public API.
